Convert background job latency to nano seconds. Move to separate module.
