+++ src/scout_apm/bottle.py
@@ -5,11 +5,11 @@ from bottle import request, response
 
 import scout_apm.core
 from scout_apm.core.config import scout_config
+from scout_apm.core.queue_time import track_request_queue_time
 from scout_apm.core.tracked_request import TrackedRequest
 from scout_apm.core.web_requests import (
     create_filtered_path,
     ignore_path,
-    track_request_queue_time,
 )
