+++ src/scout_apm/falcon.py
@@ -7,11 +7,11 @@ import falcon
 
 from scout_apm.api import install
 from scout_apm.core.config import scout_config
+from scout_apm.core.queue_time import track_request_queue_time
 from scout_apm.core.tracked_request import TrackedRequest
 from scout_apm.core.web_requests import (
     create_filtered_path,
     ignore_path,
-    track_request_queue_time,
 )
 
 logger = logging.getLogger(__name__)
